AVX-32997: Move login
